seq: passing tests 4,5,6,7,8,9 - bunch of code taken from dtolnay/seq-macro
i had a bunch of ideas, but all of them were to hard for me to
implement.
the one which looked the cleanest for solving templating problem, was to
build function which returns Option<(beginning_TT, repeated_TT,
end_TT)> (where TT = TokenTree or TokenStream).
the main culprit, was the fact i dont comprehend how to corretly split
recurrsive algorithm to hold state on few different execution arms.